pyproject.toml
[project]
name = "jotsu"
version = "0.1.0"
description = "The CLI for https://api.jotsu.com"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["click", "httpx", "python-dotenv"]

[project.scripts]
jotsu = "jotsu.cli.main:cli"

[build-system]
requires = ["setuptools>=80.3"]
build-backend = "setuptools.build_meta"
